# docs/openapi.spec.yml
openapi: 3.0.0
info:
  title: Overdeer
  description: "# Introduction\nOverdeer API provides back-end for the school organizer service.\n\n# Authentication\nUsers should be registered at /api/register and then and then a token should be obtained at /api/login and used as Bearer token with every request. \n\n# Error Codes\nGeneric errors are returned as an object with key `error` which contains the message.\nSuch as `{ \"error\": \"Group not found\" }` returned with a 404 status.\n\nErrors `400` and `422` provide more specific responses, returning keys of the invalid fields with their error messages.\n\nSuch as `{ \"email\": \"Email has to be of type string\" }` with a 400 status, or `{ \"username\": \"\"Username must be between 2 and 30 characters\" }` with 422 status."
  contact: {}
  version: "1.0"
servers:
  - url: http://localhost:8080/api
    variables: {}
paths:
  /current:
    get:
      tags:
        - Auth
      summary: Current
      description: Basic JWT Authentication
      operationId: Current
      parameters:
        - name: Content-Type
          in: header
          description: ""
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: application/json
      responses:
        200:
          description: ""
          headers:
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 75
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Tue, 15 Jan 2019 11:13:55 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/CurrentUser"
              example:
                id: 5c3d01c7ffd10a69f1cc7874
                username: example
                email: example@example.com
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
  /register:
    post:
      tags:
        - Auth
      summary: Register
      description: "TODO: Add Description"
      operationId: Register
      parameters:
        - name: Content-Type
          in: header
          description: ""
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: application/json
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example: '{"username":"test","email":"test@test.com","password":"securepasswd"}'
        required: true
      responses:
        201:
          description: ""
          headers:
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 52
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Tue, 15 Jan 2019 11:15:03 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/Register-OK"
              example:
                username: example
                email: example@example.com
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
  /login:
    post:
      tags:
        - Auth
      summary: Login
      description: "TODO: Add Description"
      operationId: Login
      parameters:
        - name: Content-Type
          in: header
          description: ""
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: application/json
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example: '{"email":"test@test.com","password":"securepasswd"}'
        required: true
      responses:
        200:
          description: ""
          headers:
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 218
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Tue, 15 Jan 2019 11:15:31 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/Login-OK"
              example:
                token: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjM2RjMGI3Njg4M2JiNDM1MTk5ZGQ4YiIsInVzZXJuYW1lIjoiZXhhbXBsZSIsImlhdCI6MTU0NzU1MDkzMSwiZXhwIjoxNTQ3NTU0NTMxfQ.Rh9BjxFsKLxrjjRrqFnzeuzdmIs7-UEHUYw0m_u54Vk
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login-Invalid"
              example:
                login: Invalid user or/and password
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
...
  /groups:
    post:
      tags:
        - Groups
      summary: Create Group
      description: "TODO: Add Description"
      operationId: CreateGroup
      parameters:
        - name: Content-Type
          in: header
          description: ""
          required: true
          style: simple
          explode: false
          schema:
            type: string
            example: application/json
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
            example: '{"name":"test","description":"test","isPrivate":false}'
        required: true
      responses:
        201:
          description: ""
          headers:
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 110
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Tue, 15 Jan 2019 10:42:27 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/CreateGroup-OK"
              example:
                _id: 5c3db913b325f140b4b24865
                name: example
                slug: example
                description: example
                isPrivate: false
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteGroup-Forbidden"
              example:
                error: This name is already taken
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
    get:
      tags:
        - Groups
      summary: List Groups
      description: "TODO: Add Description"
      operationId: ListGroups
      parameters:
        - name: page
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
        - name: count
          in: query
          description: ""
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 10
      responses:
        200:
          description: ""
          headers:
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                  example: 539
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                  example: Tue, 15 Jan 2019 10:51:05 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                  example: keep-alive
          content:
            application/json; charset=utf-8:
              schema:
                $ref: "#/components/schemas/ListGroups-UnauthenticatedOK"
              example:
                groups:
                  - isPrivate: false
                    _id: 5c3d0fca2b0fd1115bf13b56
                    name: example1
                    slug: example1
                    description: example
                    owner:
                      avatar: ""
                      _id: 5c3d01c7ffd10a69f1cc7874
                      username: example
                    createdAt: 2019-01-14T22:40:10.9720000Z
                  - isPrivate: false
                    _id: 5c3d9d161c13d8132b4c6a3e
                    name: example2
                    slug: example2
                    description: example
                    owner:
                      avatar: ""
                      _id: 5c3d01c7ffd10a69f1cc7874
                      username: example
                    createdAt: 2019-01-15T08:43:02.2390000Z
                pagination:
                  next: /api/groups?page=1&count=10
                  previous: /api/groups?page=1&count=10
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
components:
  schemas:
    CurrentUser:
      title: CurrentUser
      required:
        - id
        - username
        - email
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
      example:
        id: 5c3d01c7ffd10a69f1cc7874
        username: example
        email: example@example.com
    RegisterRequest:
      title: RegisterRequest
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        username: test
        email: test@test.com
        password: securepasswd
    Register-OK:
      title: Register-OK
      required:
        - username
        - email
      type: object
      properties:
        username:
          type: string
        email:
          type: string
      example:
        username: example
        email: example@example.com
    LoginRequest:
      title: LoginRequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: test@test.com
        password: securepasswd
    Login-OK:
      title: Login-OK
      required:
        - token
      type: object
      properties:
        token:
          type: string
      example:
        token: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjVjM2RjMGI3Njg4M2JiNDM1MTk5ZGQ4YiIsInVzZXJuYW1lIjoiZXhhbXBsZSIsImlhdCI6MTU0NzU1MDkzMSwiZXhwIjoxNTQ3NTU0NTMxfQ.Rh9BjxFsKLxrjjRrqFnzeuzdmIs7-UEHUYw0m_u54Vk
    Login-Invalid:
      title: Login-Invalid
      required:
        - login
      type: object
      properties:
        login:
          type: string
      example:
        login: Invalid user or/and password
...
    CreateGroupRequest:
      title: CreateGroupRequest
      required:
        - name
        - description
        - isPrivate
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isPrivate:
          type: boolean
      example:
        name: test
        description: test
        isPrivate: false
    CreateGroup-OK:
      title: CreateGroup-OK
      required:
        - _id
        - name
        - slug
        - description
        - isPrivate
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        isPrivate:
          type: boolean
      example:
        _id: 5c3db913b325f140b4b24865
        name: example
        slug: example
        description: example
        isPrivate: false
...
    Owner:
      title: Owner
      required:
        - avatar
        - _id
        - username
      type: object
      properties:
        avatar:
          type: string
        _id:
          type: string
        username:
          type: string
      example:
        avatar: ""
        _id: 5c3d01c7ffd10a69f1cc7874
        username: example
    ListGroups-UnauthenticatedOK:
      title: ListGroups-UnauthenticatedOK
      required:
        - groups
        - pagination
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
          description: ""
        pagination:
          $ref: "#/components/schemas/Pagination"
      example:
        groups:
          - isPrivate: false
            _id: 5c3d0fca2b0fd1115bf13b56
            name: example1
            slug: example1
            description: example
            owner:
              avatar: ""
              _id: 5c3d01c7ffd10a69f1cc7874
              username: example
            createdAt: 2019-01-14T22:40:10.9720000Z
          - isPrivate: false
            _id: 5c3d9d161c13d8132b4c6a3e
            name: example2
            slug: example2
            description: example
            owner:
              avatar: ""
              _id: 5c3d01c7ffd10a69f1cc7874
              username: example
            createdAt: 2019-01-15T08:43:02.2390000Z
        pagination:
          next: /api/groups?page=1&count=10
          previous: /api/groups?page=1&count=10
    Group:
      title: Group
      required:
        - isPrivate
        - _id
        - name
        - slug
        - description
        - owner
        - createdAt
      type: object
      properties:
        isPrivate:
          type: boolean
          example: false
        _id:
          type: string
          example: 5c3d0fca2b0fd1115bf13b56
        name:
          type: string
          example: example1
        slug:
          type: string
          example: example1
        description:
          type: string
          example: example
        owner:
          $ref: "#/components/schemas/Owner"
        createdAt:
          type: string
          example: 1/14/2019 10:40:10 PM
    Pagination:
      title: Pagination
      required:
        - next
        - previous
      type: object
      properties:
        next:
          type: string
        previous:
          type: string
...
    DeleteGroup-Forbidden:
      title: DeleteGroup-Forbidden
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: Insufficient permissions to delete this group
...